dice now responsible for annotating on map
